Add sourcing function for s3 keys
We need to pull a list of keys on each bucket with the same constraint
(prefix).
The s3 prefix has two components : one part specific to the bucket we
use, one specific to the set of keys we target (by result type / by
date)
We need the common part to calculate the set difference and copy the
result to the destination bucket.
Also, request are filtered out.