Remove the --days parameters
The naming of our documents is not consistent, so limiting by date using
only prefixes is non-trivial (if possible).
Limiting by filtering inside python does not bring much benefits, since
we already do the set difference: in both case we would the listing
would not be "prefiltered" by s3 prefix.